Support for the Throwable interface
The PH7_CLASS_THROWABLE definition is used as a class flag.

The Throwable interface is now available, where all classes that
implement it and interfaces that extend it will be throwable.

When a class is thrown, its flag is checked to determine whether it can
be thrown like an exception.

This throwable interface is fully supported, however, the only
disadvantage is that the evaluator does not check whether a class
extends the Exception class.