Method return type partially supported
This only applies to methods and not functions!

It works with interface or abstract methods just as well as class
methods.

Also, the return types are not enforced but only interpreted in terms of
syntax.

Method return types using a keyword and colon (like PHP 7) can now be
parsed by the interpreter.

The accepted types are: array, bool, int and string.

Function return types, enforced return type, and float or double return
type will be available later.